changes for teh front end for the different pages, had to movce files etc
