+++ linguistic_dna/front-end/pages/st_audiorec/frontend/.prettierrc
