+++ linguistic_dna/front-end/pages/st_audiorec/frontend/tsconfig.json
